hortening the file names (HighDynamicRange -> HDR)
